requirements.txt
Django==4.0
Pillow==8.4.0
sorl-thumbnail
djangorestframework
djangorestframework-simplejwt
